Update the AI to make it a bit smarter
This is quite a big update.
First I changed models to gemini 2.0 flash.
I implemented Bloom Taxonomy within the generated learning modules, and utilized that for quiz question generation.
There were some really stupid quiz questions being generated along with multiple questions that were the same, so I fixed that as well.
The knowledge base and learning modules are significantly more robust now as well.
I also fixed a couple of smaller issues such as the login button not recognizing when "Enter" is pressed and needing to be clicked twice. The same fix was applied to the Generator too.